Fix compiler warnings for C23 compatibility and type safety
- Fix main() return type: changed from void to int in viola/main.c
- Add proper function prototypes for deprecated non-prototype warnings:
  * initViola() in viola.h
  * findObject() in obj.h (changed parameter from int to long)
  * zoom(), smooth(), freeImage(), freeImageData() in libIMG/image.h
  * dupString() in libIMG/image.h
  * clearPacket() in viola/packet.h
  * All STG library functions in libStyle/libstg.h
- Fix implicit type conversions:
  * Float precision: use 100.0f instead of 100.0 in zoom.c
  * Color distance calculation: change unsigned long to int in send.c
    (fixes incorrect abs() usage on unsigned values)
  * Sign conversion: explicit casts in cl_client.c
- Fix pointer type mismatches:
  * Add cast for lfree(yptr) in rotate.c (byte** to byte*)
- Update format strings to match actual pointer types in libstg.c

All changes maintain backward compatibility while ensuring C23 compliance
and eliminating implicit conversions that could cause bugs.

+++ src/libIMG/send.c
@@ -373,7 +373,7 @@ imageToXImage: XAllocColor failed on a TrueColor/Directcolor visual\n");
                     for (i = 0; i < image->rgb.used; i++) {
                         /*	    if (cols[i]==0xffff) { */ /* an unallocated pixel ??*/
                         int j, d, mdist, close;
-                        unsigned long r, g, b;
+                        int r, g, b;
 
                         mdist = 100000;
                         close = -1;
@@ -381,8 +381,8 @@ imageToXImage: XAllocColor failed on a TrueColor/Directcolor visual\n");
                         g = *(image->rgb.green + i);
                         b = *(image->rgb.blue + i);
                         for (j = 0; j < dc; j++) {
-                            d = abs(r - (ctab[j].red)) + abs(g - (ctab[j].green)) +
-                                abs(b - (ctab[j].blue));
+                            d = abs(r - (int)(ctab[j].red)) + abs(g - (int)(ctab[j].green)) +
+                                abs(b - (int)(ctab[j].blue));
                             /*
                                           d = abs(r - (ctab[j].red>>8)) +
                                             abs(g - (ctab[j].green>>8)) +
